Abstracted Propellant and fixed bugs with integration code
Propellant now is a superclass. The latest implementation was renamed to
SteadyStatePropellant while the earlier implementation was readded as
EmpericalPropellant. The procedure for using a propellant model has also
changed. The model must be given the Kn, and it does all the
calculations. Then Pc and Br can be requested. 
Fixed a bug with the trapezoidal integration code in Simulation summary.